ease the process of requiring files for tests
