+++ lib/black_and_white/helpers/active_record/utils.rb
@@ -1,0 +1,21 @@
+module BlackAndWhite
+  module Helpers
+    module ActiveRecord
+      module Utils
+        extend self
+
+        def active_record_4?
+          ::ActiveRecord::VERSION::MAJOR == 4
+        end
+
+        def active_record_5?
+          ::ActiveRecord::VERSION::MAJOR == 5
+        end
+
+        def active_record_3?
+          ::ActiveRecord::VERSION::MAJOR == 3
+        end
+      end
+    end
+  end
+end
