--- lib/black_and_white/helpers/utils.rb
@@ -1,19 +1,0 @@
-module BlackAndWhite
-  module Helpers
-    module Utils
-      extend self
-
-      def active_record_4?
-        ::ActiveRecord::VERSION::MAJOR == 4
-      end
-
-      def active_record_5?
-        ::ActiveRecord::VERSION::MAJOR == 5
-      end
-
-      def active_record_3?
-        ::ActiveRecord::VERSION::MAJOR == 3
-      end
-    end
-  end
-end
