+++ lib/black_and_white/hooks.rb
@@ -1,14 +1,17 @@
 module BlackAndWhite
   class Hooks
     def self.init
+      require "black_and_white"
       if defined?(::ActiveRecord)
-        Helpers::Broker.register(:active_record)
-
+        BlackAndWhite::Broker.register(:active_record)
+        require "black_and_white/broker"
+        require "black_and_white/active_record"
         require 'black_and_white/active_record/test'
-        require 'black_and_white/active_record'
+        require "black_and_white/active_record/error"
+        require "black_and_white/helpers/active_record/utils"
+        require "black_and_white/helpers/active_record/database"
       elsif defined?(::Mongoid)
-        Helpers::Broker.register(:mongoid)
-
+        BlackAndWhite::Broker.register(:mongoid)
         require 'black_and_white/mongoid/test'
         require 'black_and_white/mongoid'
       end
