+++ spec/support/active_record/database.rb
@@ -5,16 +5,15 @@ db_config = YAML.load_file(File.expand_path('../database_config.yml', __FILE__))
 
 ActiveRecord::Base.configurations = db_config
 config = ActiveRecord::Base.configurations[db_name]
-
 begin
-  if BlackAndWhite::Helpers::Utils.active_record_4? ||
-     BlackAndWhite::Helpers::Utils.active_record_5?
+  if BlackAndWhite::Helpers::ActiveRecord::Utils.active_record_4? ||
+     BlackAndWhite::Helpers::ActiveRecord::Utils.active_record_5?
     ActiveRecord::Base.establish_connection(db_name.to_sym)
   else
     ActiveRecord::Base.establish_connection(db_name)
   end
   ActiveRecord::Base.connection
-rescue
+rescue Exception => e
   case db_name
   when 'mysql'
     ActiveRecord::Base.establish_connection(config.merge('database' => nil))
